:recycle: Refactor some components so that they include styles in the same folder
