+++ src/components/DarkModeToggle/DarkModeToggle.scss
@@ -1,3 +1,5 @@
+@import '../../styles/components/toggle';
+
 .DarkModeToggle {
   @extend .toggle;
   background: #636b78;
