+++ src/components/Footer/Footer.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { siteConfig } from '~/config/site.config';
+import './Footer.scss';
 
 export const Footer: React.FC = () => (
   <div className="Footer">
