+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { Footer } from '~/components/Footer';
-import { Navbar } from '~/components/Navbar';
+import { Footer } from '~/components/Footer/Footer';
+import { Navbar } from '~/components/Navbar/Navbar';
 import '~/styles/index.scss';
 
 interface LayoutProps {
