+++ src/components/Navbar/Navbar.scss
@@ -1,3 +1,6 @@
+@import '../../styles/tools';
+@import '../../styles/utilities';
+
 .Navbar {
   display: flex;
   flex-direction: column;
