+++ src/components/PostAuthor/PostAuthor.scss
@@ -1,3 +1,5 @@
+@import '../../styles/tools';
+
 .PostAuthor {
   display: flex;
   justify-content: flex-start;
