+++ src/components/PostPreview/PostPreview.scss
