+++ src/components/ProfileHeader/ProfileHeader.scss
@@ -1,3 +1,5 @@
+@import '../../styles/tools';
+
 .ProfileHeader {
   display: flex;
   flex-direction: column;
