+++ src/styles/components/index.scss
@@ -2,14 +2,5 @@
 @import 'button';
 @import 'hero';
 @import 'toggle';
-@import 'tooltip';
 @import 'social-link';
 @import 'filename';
-// React
-@import 'DarkModeToggle';
-@import 'Footer';
-@import 'MainTools';
-@import 'Navbar';
-@import 'PostAuthor';
-@import 'PostPreview';
-@import 'ProfileHeader';
