+++ src/styles/tools/_mixins.scss
@@ -1,3 +1,5 @@
+@import '../settings';
+
 @mixin small-breakpoint {
   @media (min-width: #{$mobile}) {
     @content;
